renamed the folder to 'srp-quiz-modal'
