+++ ui/components/app/srp-quiz-modal/QuizContent/QuizContent.tsx
