+++ ui/components/app/srp-quiz-modal/QuizContent/index.ts
