+++ ui/components/app/srp-quiz-modal/SRPQuiz/SRPQuiz.stories.js
